File-per-class, formatting code.
