+++ src/ExtendedXmlSerializer/ConverterModel/Collections/ArrayReader.cs
@@ -21,12 +21,32 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
+using System.Collections;
+using ExtendedXmlSerialization.ConverterModel.Properties;
+using ExtendedXmlSerialization.ConverterModel.Xml;
+using ExtendedXmlSerialization.Core;
+using ExtendedXmlSerialization.TypeModel;
+
 namespace ExtendedXmlSerialization.ConverterModel.Collections
 {
-/*
-	class CollectionItem : Container
+	class ArrayReader : CollectionReader
 	{
-		public CollectionItem(IWriter element, IConverter body) : base(element, body) {}
+		readonly ITypeProperty _property;
+
+		public ArrayReader(IConverter item) : this(item, AddDelegates.Default, ItemTypeProperty.Default) {}
+
+		public ArrayReader(IConverter item, IAddDelegates add, ITypeProperty property)
+			: base(ArrayActivator.Default, item, add)
+		{
+			_property = property;
+		}
+
+		public override object Get(IXmlReader parameter)
+		{
+			var itemType = _property.Get(parameter);
+			var list = base.Get(parameter).AsValid<ArrayList>();
+			var result = list.ToArray(itemType.AsType());
+			return result;
+		}
 	}
-*/
 }
