+++ src/ExtendedXmlSerializer/ConverterModel/Collections/CollectionReader.cs
@@ -24,39 +24,12 @@
 using System.Collections;
 using System.Reflection;
 using ExtendedXmlSerialization.ConverterModel.Converters;
-using ExtendedXmlSerialization.ConverterModel.Properties;
 using ExtendedXmlSerialization.ConverterModel.Xml;
 using ExtendedXmlSerialization.Core;
 using ExtendedXmlSerialization.TypeModel;
 
 namespace ExtendedXmlSerialization.ConverterModel.Collections
 {
-	class ArrayReader : CollectionReader
-	{
-		readonly ITypeProperty _property;
-
-		public ArrayReader(IConverter item) : this(item, AddDelegates.Default, ItemTypeProperty.Default) {}
-
-		public ArrayReader(IConverter item, IAddDelegates add, ITypeProperty property) : base(ArrayActivator.Default, item, add)
-		{
-			_property = property;
-		}
-
-		public override object Get(IXmlReader parameter)
-		{
-			var itemType = _property.Get(parameter);
-			var list = base.Get(parameter).AsValid<ArrayList>();
-			var result = list.ToArray(itemType.AsType());
-			return result;
-		}
-	}
-
-	class ArrayActivator : DelegatedFixedReader
-	{
-		public static ArrayActivator Default { get; } = new ArrayActivator();
-		ArrayActivator() : base(() => new ArrayList()) {}
-	}
-
 	class CollectionReader : DecoratedReader
 	{
 		readonly IConverter _item;
