+++ src/ExtendedXmlSerializer/ConverterModel/Properties/ItemTypeProperty.cs
