+++ src/ExtendedXmlSerializer/ConverterModel/Xml/IXmlReader.cs
@@ -29,7 +29,8 @@ using ExtendedXmlSerialization.Core.Sources;
 
 namespace ExtendedXmlSerialization.ConverterModel.Xml
 {
-	public interface IXmlReader : IDisposable, IDisplayAware, IParameterizedSource<string, XName>, IParameterizedSource<XName, TypeInfo>
+	public interface IXmlReader : IDisposable, IDisplayAware, IParameterizedSource<string, XName>,
+	                              IParameterizedSource<XName, TypeInfo>
 	{
 		TypeInfo Classification();
